+++ juriscraper/opinions/united_states/federal_special/fisc.py
@@ -13,7 +13,7 @@ class Site(OpinionSiteLinear):
         self.do_backscrape = False
         self.should_have_results = True
 
-    def _process_html(self) -> None:
+    async def _process_html(self) -> None:
         """Parse HTML into case objects
 
         :return None
@@ -44,8 +44,8 @@ class Site(OpinionSiteLinear):
                 # Perform download since we are not yielding a site for each
                 # element in the backscrape iterable
                 self.url = urljoin(self.url, next_page[0])
-                self.html = self._download()
-                self._process_html()
+                self.html = await self._download()
+                await self._process_html()
 
     def skip_record(self, docket: str, filename: str) -> bool:
         """Check if a record belongs to 'FISCR'