+++ juriscraper/opinions/united_states/state/iowactapp.py
@@ -13,13 +13,13 @@ class Site(iowa.Site):
         self.court_id = self.__module__
         self.url = "https://www.iowacourts.gov/iowa-courts/court-of-appeals/court-of-appeals-court-opinions/"
 
-    def _download_backwards(self, _):
+    async def _download_backwards(self, _):
         """Overriding this method from parent iowa class because the
         site link/page structure for archived Court of Appeals opinions
         is different than that of Archived Supreme Court opinions
         """
         self.archive = True
-        landing_page_html = self._download()
+        landing_page_html = await self._download()
         path_filter_class = 'contains(./@class, "nav-link")'
         path_filter_text = (
             'contains(./text(), "Archived Court of Appeals Opinions")'