+++ juriscraper/pacer/docket_history_report.py
@@ -90,7 +90,7 @@ class DocketHistoryReport(DocketReport):
         self._metadata = data
         return data
 
-    def query(
+    async def query(
         self,
         pacer_case_id,
         query_type="History",
@@ -129,7 +129,7 @@ class DocketHistoryReport(DocketReport):
             "Getting nonce for docket history report with "
             "pacer_case_id: %s" % pacer_case_id
         )
-        r = self.session.get(f"{self.url}?{pacer_case_id}")
+        r = await self.session.get(f"{self.url}?{pacer_case_id}")
         nonce = get_nonce_from_form(r)
 
         query_params = {
@@ -144,7 +144,7 @@ class DocketHistoryReport(DocketReport):
             "params %s and nonce %s" % (pacer_case_id, query_params, nonce)
         )
 
-        self.response = self.session.post(
+        self.response = await self.session.post(
             f"{self.url}?{nonce}", data=query_params
         )
         self.parse()