+++ tests/network/test_PacerAcmsAttachmentPageTest.py
@@ -5,12 +5,12 @@ from juriscraper.pacer import ACMSAttachmentPage
 from tests.network import SKIP_IF_NO_PACER_LOGIN, get_pacer_session
 
 
-class AcmsAttachmentPageTest(unittest.TestCase):
+class AcmsAttachmentPageTest(unittest.IsolatedAsyncioTestCase):
     """A test of basic info for the Case Query"""
 
-    def setUp(self):
+    async def asyncSetUp(self):
         self.session = get_pacer_session()
-        self.session.login()
+        await self.session.login()
         self.report = ACMSAttachmentPage("ca2", self.session)
         self.pacer_case_id = "70c4875d-f112-48e5-ad41-5e6d403ca9cd"  # Ogunbekun v. Town of Brighton
         self.docket_doc_id = (
@@ -18,9 +18,9 @@ class AcmsAttachmentPageTest(unittest.TestCase):
         )
 
     @SKIP_IF_NO_PACER_LOGIN
-    def test_queries(self):
+    async def test_queries(self):
         """Do a variety of queries work?"""
-        self.report.query(self.pacer_case_id, self.docket_doc_id)
+        await self.report.query(self.pacer_case_id, self.docket_doc_id)
         data = self.report.data
 
         self.assertEqual(